+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Extend backoff time period for pl001.
+
+### Changed
+
 - Shift organization to `conformance-testing`.
 
 ## [15.0.3] - 2021-07-16